requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
uvicorn
python-dotenv
requests-oauthlib
Pillow
requests
boto3